requirements.txt
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.60.2
gmpy2==2.2.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.1
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pip==26.0.1
pycparser==2.23
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
setuptools==80.9.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.3
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
wheel==0.45.1
zipp==3.23.0
